requirements.txt
pytorch==1.12.1
transformers==4.24.0
tensorboard==2.10.0
chardet==4.0.0
nltk==3.7
functools
wikipedia2vec